+++ lib/restclient/exceptions.rb
@@ -34,7 +34,11 @@ module RestClient
     EXCEPTIONS_MAP = {}
   end
 
-  {304 => 'Not Modified',
+  {300 => 'Multiple Choices',
+   301 => 'Moved Permanently',
+   302 => 'Found',
+   303 => 'See Other',
+   304 => 'Not Modified',
    305 => 'Use Proxy',
    400 => 'Bad Request',
    401 => 'Unauthorized',