+++ history.md
@@ -1,6 +1,7 @@
 # 1.2.1
 
 - cleaner log API, add a warning for some cases but should be compatible
+- accept multiple "Set-Cookie" headers, see http://www.ietf.org/rfc/rfc2109.txt (patch provided by Cyril Rohr)
 
 # 1.2.0