+++ history.md
@@ -1,9 +1,10 @@
 # 1.2.1
 
+- a block can be used to process a request's result, this enable to handle custom error codes or paththrought (design by Cyril Rohr)
 - cleaner log API, add a warning for some cases but should be compatible
 - accept multiple "Set-Cookie" headers, see http://www.ietf.org/rfc/rfc2109.txt (patch provided by Cyril Rohr)
 - remove "Content-Length" and "Content-Type" headers when following a redirection (patch provided by haarts)
-- all http error codes has now a corresponding exception class and all of them contain the Reponse
+- all http error codes have now a corresponding exception class and all of them contain the Reponse
 
 # 1.2.0